refact(core): inductive types

+++ saki-concrete/src/main/scala/saki/concrete/Visitor.scala
@@ -469,8 +469,8 @@ class Visitor extends SakiBaseVisitor[SyntaxTree[?] | Seq[SyntaxTree[?]]] {
     val value = ctx.literal match {
       case _: LiteralBoolContext => LiteralValue.BoolValue(valueString.toBoolean)
       case _: LiteralCharContext => LiteralValue.RuneValue(valueString.charAt(1))
-      case _: LiteralFloatContext => LiteralValue.FloatValue(valueString.toFloat)
-      case _: LiteralIntContext => LiteralValue.IntValue(valueString.toInt)
+      case _: LiteralFloatContext => LiteralValue.FloatValue(valueString.toDouble)
+      case _: LiteralIntContext => LiteralValue.IntValue(valueString.toLong)
       case _: LiteralRegularStringContext => LiteralValue.StringValue(valueString.substring(1, valueString.length - 1))
       case _: LiteralRawStringContext => {
         LiteralValue.StringValue(valueString.stripPrefix("#").stripPrefix("{").stripSuffix("}"))
